feat: support Wiimote
The ndx factor is the difficult part. But other than that I could make
this JS work for Wiimote on Firefox and Chromium.

Firefox registers a single Wiimote as different gamepads which mostly
works with your implementation.

In Chromium I have only 7 buttons registered which fails the DPad. I
could not make the cross (analog DPad, if you will) work.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>